Moved Bot class to separate file.
